// src/app/demo/models/modelsbienraiz/TipoDocumentoViewModel.ts
export class tipodocumento{
    tido_Id?: number;
    tido_Descripcion?: string;
    usua_Creacion?: number;
    tido_FechaCreacion?: string;
    usua_Mofificacion?: number;
    tido_FechaModificacion?: string;
    tido_Estado?: string;
    usuaCreacion?: string;
    usuaModificacion?:string;
    codigo?: string;
}